abstract supabase from Brain class
